requirements.txt
Flask
Pillow
scipy
numpy
scikit-learn
